Refactor large files by extracting components and services
Reduced file sizes significantly by extracting reusable components:
- home_page.dart: 1587 → 1417 lines (-170 lines, -10.7%)
- note_editor_page.dart: 846 → 593 lines (-253 lines, -29.9%)

New files created:
- lib/services/note_operations_service.dart: Centralized note CRUD operations
- lib/widgets/home_page/note_dialogs.dart: Dialog widgets for home page
- lib/widgets/note_editor/editor_dialogs.dart: Dialog widgets for editor
- lib/widgets/note_editor/category_chip.dart: Reusable category chip widget

Total lines reduced: 423 lines
This improves code maintainability, readability, and reusability.
